Added custom hooks for the Flashcard Set page
