+++ src/pages/FlashcardSet/FlashcardSet/hooks/useFlipMyFlashcard.tsx
@@ -1,0 +1,10 @@
+import { useState } from "react";
+
+export const useFlipMyFlashcard = () => {
+  const [isFlipped, setIsFlipped] = useState(false);
+
+  const handleFlip = () => {
+    setIsFlipped(!isFlipped);
+  };
+  return { isFlipped, handleFlip };
+};
