+++ src/pages/MyFlashcards/FlashcardSetEdit/EditFlashcardSet.tsx
@@ -23,12 +23,12 @@ import {
 
 import { DeleteIcon } from "@chakra-ui/icons";
 import ButtonPrimary from "../../../shared/ui/components/ButtonPrimary";
-import { useEditFlashcardSet } from "../../../shared/hooks/useEditFlashcardSet";
-import { useAddRemoveFlashcard } from "../../../shared/hooks/useAddRemoveFlashcard";
-import { useSaveEditedSet } from "../../../shared/hooks/useSaveEditedSet";
+import { useEditFlashcardSet } from "../../../shared/hooks/EditFlashcardSet/useEditFlashcardSet";
+import { useAddRemoveFlashcard } from "../../../shared/hooks/EditFlashcardSet/useAddRemoveFlashcard";
+import { useSaveEditedSet } from "../../../shared/hooks/EditFlashcardSet/useSaveEditedSet";
 import { EditFlashcardSetProps } from "../../../shared/types/typesEditFlashcardSet";
 import { Flashcard } from "../../../shared/types/typesFlashcard";
-import { useValidation } from "../../../shared/hooks/useValidation";
+import { useValidation } from "../../../shared/hooks/EditFlashcardSet/useValidation";
 
 export default function EditFlashcardSet({
   isOpen,
